pyproject.toml
[project]
name = "vectorize-ecg"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers",
    "pillow",
    "qdrant-client",
    "tqdm"
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu124" },
]
torchvision = [
    { index = "pytorch-cu124" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
